pyproject.toml
[project]
name = "o-pathmanager"
version = "0.1.0"
description = "pathmanager API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ['pyyaml']
